+++ CHANGELOG.md
@@ -9,7 +9,8 @@ The format is based on **[Keep a Changelog](https://keepachangelog.com/en/1.1.0/
 ## [Unreleased]
 
 - **Added**
-  - (placeholder)
+  - Reset for useWebSpeechEngine added
+  - track add for dispose events
 
 - **Changed**
   - (placeholder)